Updated wallet balancev2 and added inheritance in appwallet model
Updated wallet dashboard
removed walletinfo model
fixed txhistorty issue of parsing double to decimal
Code refactoring and name updates
updated totalusd value and totalwalletusd value calc
updated many double conversion to decimal
replaced truncatefunc to updated decimallimiter func
Updated buysellVM calculations
updated orderbook model and view
updated add gas view and VM
updated routes arguments
updated receive address view
updated trade service

+++ lib/screens/exchange/trade/orderbook/orderbook_viewmodel.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:decimal/decimal.dart';
 import 'package:exchangilymobileapp/logger.dart';
 import 'package:exchangilymobileapp/models/shared/pair_decimal_config_model.dart';
 import 'package:exchangilymobileapp/screens/exchange/trade/orderbook/orderbook_model.dart';
@@ -85,7 +86,7 @@ class OrderbookViewModel extends StreamViewModel {
 /*----------------------------------------------------------------------
                     Set price and quantity
 ----------------------------------------------------------------------*/
-  fillTextFields(double p, double q) {
+  fillTextFields(Decimal p, Decimal q) {
     tradeService.setPriceQuantityValues(p, q);
     log.i('fillTextFields $p--$q');
   }
